deprecated PrismTask.target, replaced with prism.target.target; added support for multiple targets and target iteration
