+++ prism/tests/integration/test_projects/008_targets/modules/txt.py
@@ -15,8 +15,8 @@ Table of Contents:
 #############
 
 import prism_project
-from prism.task import PrismTask       # Not necessary; prism infrastructure automatically imported on the back-end
-import prism.target as PrismTarget     # Not necessary; prism infrastructure automatically imported on the back-end
+from prism.task import PrismTask
+from prism.target import target, Txt
 
 
 ######################
@@ -26,7 +26,7 @@ import prism.target as PrismTarget     # Not necessary; prism infrastructure aut
 class Module02(PrismTask):
 
     ## Run    
-    @PrismTask.target(type=PrismTarget.Txt, loc=f'{prism_project.OUTPUT}/module02.txt')
+    @target(type=Txt, loc=f'{prism_project.OUTPUT}/target_txt.txt')
     def run(self, psm):
         string = "Hello, world!"
         return string
