+++ prism/tests/integration/test_projects/008_targets/output/target_csv_mult_df1.csv
@@ -1,0 +1,4 @@
+col1,col2,col3,col4,col5,col6
+col1_value1,col2_value1,col3_value1,col4_value1,col5_value1,col6_value1
+col1_value2,col2_value2,col3_value2,col4_value2,col5_value2,col6_value2
+col1_value3,col2_value3,col3_value3,col4_value3,col5_value3,col6_value3
