+++ prism/tests/integration/test_projects/008_targets/output/target_csv_mult_df2.csv
@@ -1,0 +1,4 @@
+colA,colB,colC,colD,colE,colF
+colA_value1,colB_value1,colC_value1,colD_value1,colE_value1,colF_value1
+colA_value2,colB_value2,colC_value2,colD_value2,colE_value2,colF_value2
+colA_value3,colB_value3,colC_value3,colD_value3,colE_value3,colF_value3
